:lipstick: contributions数の取得のためのパース作業をWebサービス側に移動させた
それに共ない不要なメソッド、クラス等を削除した。
